+++ core/src/routes/(editor)/editor/editortree.svelte
@@ -10,7 +10,7 @@
 
 
 {#each component?.children ?? [] as child}
-    <svelte:component this={JsonOfModules[child.type]} data={child.data} uuid={child.uuid}>
+    <svelte:component this={JsonOfModules[child.type]} data={child.data} uuid={child.uuid}  selected={child.selected}>
         <svelte:self component={child} JsonOfModules={JsonOfModules} >
             <!-- <slot component={child} />  -->